pulled route logic into a controller
